Rename types.ts -> types.d.ts
