+++ compiler/mod.ts
@@ -2,7 +2,7 @@ import { join } from 'https://deno.land/std@0.100.0/path/mod.ts'
 import { ensureDir } from 'https://deno.land/std@0.100.0/fs/ensure_dir.ts'
 import { existsFile } from '../shared/fs.ts'
 import { Measure } from '../shared/log.ts'
-import type { ImportMap } from '../types.ts'
+import type { ImportMap } from '../types.d.ts'
 import { getDenoDir } from '../server/helper.ts'
 import { checksum } from './dist/checksum.js'
 import init, { parseExportNamesSync, stripSsrCodeSync, transformSync } from './dist/compiler.js'
