+++ framework/react/context.ts
@@ -1,5 +1,5 @@
 import { createContext, ReactNode } from 'https://esm.sh/react@17.0.2'
-import type { RouterURL } from '../../types.ts'
+import type { RouterURL } from '../../types.d.ts'
 import { createBlankRouterURL } from '../core/routing.ts'
 import { createNamedContext } from './helper.ts'
 import type { RendererStore } from './renderer.ts'
