+++ import_map.json
@@ -1,9 +1,9 @@
 {
   "imports": {
+    "aleph/": "./",
+    "aleph/web": "./framework/core/mod.ts",
+    "aleph/react": "./framework/react/mod.ts",
     "std/": "https://deno.land/std@0.100.0/",
-    "aleph/": "https://deno.land/x/aleph@v0.3.0-alpha.33/",
-    "aleph/web": "https://deno.land/x/aleph@v0.3.0-alpha.33/framework/core/mod.ts",
-    "aleph/react": "https://deno.land/x/aleph@v0.3.0-alpha.33/framework/react/mod.ts",
     "react": "https://esm.sh/react@17.0.2",
     "react-dom": "https://esm.sh/react-dom@17.0.2"
   }
