+++ server/renderer.ts
@@ -1,7 +1,7 @@
 import { basename, dirname } from 'https://deno.land/std@0.100.0/path/mod.ts'
 import log from '../shared/log.ts'
 import util from '../shared/util.ts'
-import type { RouterURL } from '../types.ts'
+import type { RouterURL } from '../types.d.ts'
 import type { Aleph, Module } from './aleph.ts'
 
 export type SSRData = {
