+++ server/server.ts
@@ -4,7 +4,7 @@ import { resolveURL } from '../framework/core/routing.ts'
 import { existsFile } from '../shared/fs.ts'
 import log from '../shared/log.ts'
 import util from '../shared/util.ts'
-import { APIContext } from '../types.ts'
+import { APIContext } from '../types.d.ts'
 import { Aleph } from './aleph.ts'
 import compress from './compress.ts'
 import { getContentType } from './mime.ts'
