+++ version.ts
@@ -18,7 +18,7 @@ export async function prepublish(version: string): Promise<boolean> {
     const importMap = JSON.parse(data)
     Object.assign(importMap.imports, {
       'aleph/': `https://deno.land/x/aleph@v${version}/`,
-      'aleph/types': `https://deno.land/x/aleph@v${version}/types.ts`,
+      'aleph/types': `https://deno.land/x/aleph@v${version}/types.d.ts`,
       'aleph/web': `https://deno.land/x/aleph@v${version}/framework/core/mod.ts`,
       'aleph/react': `https://deno.land/x/aleph@v${version}/framework/react/mod.ts`,
       'react': `https://esm.sh/react@${defaultReactVersion}`,
